MarkoPolo Generator
Added a script which allows to generate up to 100 lines of text. On default, it generates numbers from 1-100. But every number that can be divided by 3 will be replaced by Marko, every number that can be divided by 5 will be replaced with Polo, and every number that can be divided by both 3 or 5 (in other words, numbers that can be divided by 15) will generate MarkoPolo.